Generate assets do a separate folder.
Adds a asset generate + watch script.
This cleans up the mixing of source and output css and js in the static
folders, and prevents the problem with docker volumes also removing
the ability update those source files.
This also enables us to later upload this volume to some other location.