clean up asset loading and configuration loading.
assets are now compiled using esbuild (fast!). the python app interprets
a meta file written by the script to see what assets are compiled, and
which filenames are generated. the output filename also has a hash in
it, which is good for cache invalidation (better than the old timestamp
there was). there's also the option to watch this metadata file for
changes, so that uchan can automatically reload what assets it provides
during development. this works without delay, because esbuild is so
fast.

while we're using a bundler/minifier anyway now, I've converted the
styles to sass.

additionally the assets were consolidated a bit, and there's now a
sitewide javascript file, the extra javascript file (which uses more
"advanced" javascript code), the global style, the themes, and the mod
style. there are neatly configured in the esbuild script, and
interpreted from the python side.

and finally, uchan can host local assets using werkzeug. this is more
ment as a fallback / development mode, since assets should be hosted by
a reverse proxy (nginx) or cdn. later this can be made to only work in
debug mode, once the hosting part of uchan is reworked.

the additional big thing in this commit are the changes on how uchan is
configured. this should've been in a separate commit, but it got
intertwined with the asset reworking, and at this stage no one is
bothered by that anyway.

settings now uses BaseSettings from pydantic. this has the major
advantage of typechecking our settings properly, and it loads the
settings from env vars. this is very useful with - for example -
kubernetes.