Rework Dockerfile and add Helm chart
This replaces the docker-compose stack that existed before.

The Helm chart is a combination of uwsgi + celery worker + postgres +
rabbitmq + varnish. If varnish is enabled (the default), then it
requires the excelleng varnish-operator by IBM. How µchan is installed,
will later be documented in a more clear way.

The Dockerfile builds both the front-end static files, and embed a
virtualenv (with Poetry) with all dependencies.

This also automates the setup, which is done automatically on startup if
nothing has been configured yet. The replacement is called using a new cli interface.